requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
ipython==8.18.1
